+++ utils.h
@@ -6,6 +6,6 @@
 #include <stdlib.h>
 
 uint8_t     mx(uint8_t m1, uint8_t m2);
-int         my_pow(int x, size_t n);
+unsigned long long int         my_pow(int x, size_t n);
 
 #endif
